PR feedback
Move tests for free text annotation to its separate file
